requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
supabase==2.4.3
pydantic==2.7.1
python-multipart==0.0.9
PyJWT==2.8.0
httpx==0.27.0
reportlab==4.2.0
psycopg2-binary==2.9.9
pytest==8.2.0
pytest-mock==3.14.0
